refactor: group CUDA MoE kernels
