+++ lmdeploy/pytorch/kernels/cuda/__init__.py
@@ -4,11 +4,11 @@ from .apply_rotary_pos_emb import apply_rotary_pos_emb
 from .fill_kv_cache import fill_kv_cache
 from .flashattention import flash_attn_varlen_func
 from .flatten_kv_cache import flatten_kv_cache
-from .fused_moe import fused_moe
+from .moe.fused_moe import fused_moe
+from .moe.w8a8 import fused_moe_w8a8
 from .multinomial_sampling import multinomial_sampling
 from .pagedattention import flash_attn_with_kvcache
 from .rms_norm import rms_norm
-from .w8a8_fused_moe import fused_moe_w8a8
 from .w8a8_triton_kernels import matmul_kernel_dynamic_quant, per_token_quant_int8, rms_norm_dynamic_quant
 
 __all__ = [
