+++ lmdeploy/pytorch/kernels/cuda/moe/fused_moe.py
@@ -6,7 +6,7 @@ import torch
 import triton
 import triton.language as tl
 
-from .activation import silu_and_mul
+from ..activation import silu_and_mul
 
 
 def get_cuda_autotune_config():
